third_party/blink/web_tests/wpt_internal/overscroll-gestures/overscroll-container-overlay.html

<!DOCTYPE html>
<html >
<title>Overlay overscroll container</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  html, body {
    margin: 0;
  }
  #container {
    width: 200px;
    height: 200px;
    overflow: auto;
  }
  #menu {
    width: 100%;
    height: 100%;
    left: 100%; /* Overscrolls to the right of the container. */
  }
  #content {
    height: 600px;
  }
</style>
<div id="container" overscrollcontainer="overlay">
  <div id="menu"></div>
  <div id="content"></div>
</div>
<button id=button command="toggle-overscroll" commandfor="menu"></button>
<script>
test(() => {
  // Initially the menu is offscreen to the right.
  assert_equals(menu.getBoundingClientRect().left,
                content.getBoundingClientRect().right);

  // Bring the menu into view.
  menu.scrollIntoView();
  assert_equals(menu.getBoundingClientRect().left, 0);

  // The content remains in place under the menu.
  assert_equals(menu.getBoundingClientRect().left,
                content.getBoundingClientRect().left);
}, "overlay overscrollcontainer doesn't shift the content");
</script>
</html>
